# GitRepoMind/server/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "GitRepoMind API",
                 "description":  "AI-powered GitHub Repository Analyzer \u0026 RAG Chat",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/analyze-repo":  {
                                            "post":  {
                                                         "tags":  [
                                                                      "analyze"
                                                                  ],
                                                         "summary":  "Analyze Repository",
                                                         "description":  "Analyze a GitHub repository and prepare it for RAG queries.\n\nThis endpoint performs the following steps:\n1. Parses the GitHub repository URL\n2. Checks cache for existing analysis\n3. Fetches all files from the repository\n4. Analyzes static repository structure\n5. Chunks files into semantic units\n6. Tags chunks with metadata\n7. Generates embeddings\n8. Stores vectors in Qdrant\n9. Caches the analysis result\n\nArgs:\n    request: AnalyzeRequest with repo_url and optional branch\n\nReturns:\n    AnalyzeResponse with success status, repo_id, and analysis details\n\nRaises:\n    HTTPException: For invalid URLs, GitHub errors, or processing failures",
                                                         "operationId":  "analyze_repository_api_analyze_repo_post",
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  "@{schema=}"
                                                                                         },
                                                                             "required":  true
                                                                         },
                                                         "responses":  {
                                                                           "201":  {
                                                                                       "description":  "Successful Response",
                                                                                       "content":  "@{application/json=}"
                                                                                   },
                                                                           "422":  {
                                                                                       "description":  "Validation Error",
                                                                                       "content":  "@{application/json=}"
                                                                                   }
                                                                       }
                                                     }
                                        },
                  "/api/reindex":  {
                                       "post":  {
                                                    "tags":  [
                                                                 "analyze"
                                                             ],
                                                    "summary":  "Reindex Repository",
                                                    "description":  "Re-analyze a GitHub repository, discarding cached data.\n\nThis endpoint forces a fresh analysis of a repository:\n1. Deletes existing vectors from Qdrant\n2. Clears cached analysis\n3. Runs the full analysis pipeline\n\nArgs:\n    request: ReindexRequest with repo_url and optional branch\n\nReturns:\n    Fresh AnalyzeResponse with new analysis data\n\nRaises:\n    HTTPException: For invalid URLs or processing failures",
                                                    "operationId":  "reindex_repository_api_reindex_post",
                                                    "requestBody":  {
                                                                        "content":  {
                                                                                        "application/json":  "@{schema=}"
                                                                                    },
                                                                        "required":  true
                                                                    },
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "Successful Response",
                                                                                  "content":  "@{application/json=}"
                                                                              },
                                                                      "422":  {
                                                                                  "description":  "Validation Error",
                                                                                  "content":  "@{application/json=}"
                                                                              }
                                                                  }
                                                }
                                   },
                  "/api/analysis/{repo_id}":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "analyze"
                                                                       ],
                                                              "summary":  "Get Analysis",
                                                              "description":  "Retrieve cached analysis for a repository.\n\nArgs:\n    repo_id: Repository ID returned from analyze-repo endpoint\n\nReturns:\n    Cached AnalyzeResponse\n\nRaises:\n    HTTPException: If repository not found",
                                                              "operationId":  "get_analysis_api_analysis__repo_id__get",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "repo_id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  "@{type=string; title=Repo Id}"
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Successful Response",
                                                                                            "content":  "@{application/json=}"
                                                                                        },
                                                                                "422":  {
                                                                                            "description":  "Validation Error",
                                                                                            "content":  "@{application/json=}"
                                                                                        }
                                                                            }
                                                          }
                                              },
...
              },
...
}